requirements.txt
psycopg2-binary
python-dotenv
requests
gspread
google-auth
matplotlib
